% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis.R
\name{conterfactuals_check}
\alias{conterfactuals_check}
\title{Checks counterfactuals and plot the results.
Only to check one to one relationship  A -> B.}
\usage{
conterfactuals_check(posterior)
}
\arguments{
\item{posterior}{A number.}
}
\description{
Checks counterfactuals and plot the results.
Only to check one to one relationship  A -> B.
}
\examples{
conterfactuals_check(1)
conterfactuals_check(10)
}
