% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{feature_density}
\alias{feature_density}
\title{Plot a density graph of feature's names contained in the dataset.}
\usage{
feature_density(column.names, climateset)
}
\arguments{
\item{column.names}{Must be of type formula.}

\item{climateset}{Must be a dataframe.}
}
\description{
Plot a density graph of feature's names contained in the dataset.
}
\examples{
feature_density(1, 1)
}
