% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{feature_hist}
\alias{feature_hist}
\title{Create histograms of feature's names contained in the dataset.}
\usage{
feature_hist(column.names, climateset)
}
\arguments{
\item{column.names}{Must be of type formula.}

\item{climateset}{Must be a dataframe.}
}
\description{
Create histograms of feature's names contained in the dataset.
}
\examples{
feature_hist(1, 1)
}
