% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis.R
\name{make_forecast}
\alias{make_forecast}
\title{Make forecast on new data based on the posterior.}
\usage{
make_forecast(posterior, new_data)
}
\arguments{
\item{posterior}{A number.}

\item{new_data}{Must be a dataframe.}
}
\value{
The prediction.
}
\description{
Make forecast on new data based on the posterior.
}
\examples{
make_forecast(1, 1)
make_forecast(10, 1)
}
