% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_predictive_check.R
\name{model_check}
\alias{model_check}
\title{Check the model and display the results.}
\usage{
model_check(feature.names, climateset, mu = 0, sigma = 1, prior = TRUE)
}
\arguments{
\item{feature.names}{A list of features.}

\item{climateset}{Must be a dataframe.}

\item{mu}{A number set per defalt to 0.}

\item{sigma}{A number set per default to 1.}

\item{prior}{A boolean TRUE means prior_predictive_check is called, else Posterior_predictive_check. Per Default is its value set to TRUE.}
}
\description{
Check the model and display the results.
}
\examples{
model_check([], 1)
}
