% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis.R
\name{posterior_analysis}
\alias{posterior_analysis}
\title{Analyse the posterior}
\usage{
posterior_analysis(
  formula,
  climateset,
  verbose = FALSE,
  conterfactuals.check = FALSE
)
}
\arguments{
\item{formula}{Must be of type formula.}

\item{climateset}{Must be a dataframe.}

\item{verbose}{A boolean set per default to FALSE}

\item{conterfactuals.check}{A boolean set per default to FALSE}
}
\value{
The posterior.
}
\description{
Analyse the posterior
}
\examples{
posterior_analysis(1, 1)
posterior_analysis(10, 1)
}
