% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{ppc_display}
\alias{ppc_display}
\title{Plot a ppc density overlay the provided data.}
\usage{
ppc_display(data, yrep)
}
\arguments{
\item{data}{Must be of type dataframe.}
}
\description{
Plot a ppc density overlay the provided data.
}
\examples{
feature_density(1, 1)
}
